Address final mTLS review comments
