+++ tests/Snippets/MutualTlsReadMeSnippets.cs
@@ -43,6 +43,7 @@ public class MutualTlsReadMeSnippets
                 SslStreamCertificateContext.Create(
                     clientCertificate,
                     intermediateCertificates,
+                    // Build the local chain without revocation network lookups.
                     offline: true);
 
             SocketsHttpHandler handler = new()
